Add helper function loadRuntimeModule
The rationale for this is that this makes the code which has to do this
easier to read. Currently that's only `dmd.expressionsem`.
It is expected that the functionality to load runtime modules will be
used more in the future to simplify parts of the compiler once
this convenient way to do so is provided.